refractor: moved get_config_templates, get_modules_cbm_config, get_provider_config to preprocess.py file
